feat(mu): move cron into the server #661
